[댓글 기능] 코니 미션 제출합니다. (#157)
* feat: LoginInterceptor, LogoutInterceptor 구현 및 일부 uri 수정

- 수정된 uri: writing -> articles/new, signup -> users/new

* feat: Comment class 구현

* feat: CommentRepository 구현

* feat: User와 Article 관계 맵핑

- Article 수정 및 삭제시 권한 확인하는 로직 추가
- 세션에 LoginUser 대신 User 저장하도록 변경

* refactor: Controller의 ExceptionHandler를 GlobalExceptionHandler로 분리

* feat: Comment 작성 기능 구현

* feat: index와 article에서 댓글 개수 올바르게 표기하도록 구현

* feat: Comment 수정 기능 구현

* feat: Comment 삭제 기능 구현

* refactor: HomeController 분리 및 ArticleController uri 중복 제거

* feat: Article, Comment의 contents에 @Lob 적용 및 title 글자수 제한 수정

* feat: 모든 Entity가 생성 시각 및 최종 수정 시각을 갖도록 구현

* feat: Comment 작성 시간이 출력되도록 구현

* refactor: assembler 패키지 분리

* feat: Article, User의 생성자에서 id parameter 제거

* feat: 모든 dto를 request, response로 분리 및 dto setter 제거

* fix: Article coverUrl의 length 제한을 100에서 200으로 변경

* refactor: interceptor, configuration 패키지 분리

* test: UserTest, ArticleServiceTest, UserServiceTest 구현

- saveComment에 @Transactional 추가
- Article의 comments에 FetchType.LAZY 적용
- data.sql로 DB에 테스트 데이터 적재

* test: ArticleControllerTest, UserControllerTest 구현

* feat: application.properties 업데이트

- data.sql 삽입을 위한 spring.datasource.initialization-mode=always 추가

* refactor: uri 수정

- new, edit, delete 등 불필요한 uri 요소 제거

+++ src/main/java/techcourse/myblog/service/ArticleService.java
@@ -2,10 +2,15 @@ package techcourse.myblog.service;
 
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import techcourse.myblog.domain.Article;
-import techcourse.myblog.domain.ArticleAssembler;
-import techcourse.myblog.dto.ArticleDto;
+import techcourse.myblog.domain.*;
+import techcourse.myblog.domain.assembler.ArticleAssembler;
+import techcourse.myblog.domain.assembler.CommentAssembler;
+import techcourse.myblog.dto.ArticleRequest;
+import techcourse.myblog.dto.ArticleResponse;
+import techcourse.myblog.dto.CommentRequest;
+import techcourse.myblog.dto.CommentResponse;
 import techcourse.myblog.repository.ArticleRepository;
+import techcourse.myblog.repository.CommentRepository;
 
 import java.util.List;
 import java.util.NoSuchElementException;
@@ -13,35 +18,75 @@ import java.util.NoSuchElementException;
 @Service
 public class ArticleService {
     private final ArticleRepository articleRepository;
+    private final CommentRepository commentRepository;
 
-    public ArticleService(ArticleRepository articleRepository) {
+    public ArticleService(ArticleRepository articleRepository, CommentRepository commentRepository) {
         this.articleRepository = articleRepository;
+        this.commentRepository = commentRepository;
     }
 
-    public List<ArticleDto> getAllArticles() {
+    public List<ArticleResponse> getAllArticles() {
         return ArticleAssembler.writeDtos(articleRepository.findAll());
     }
 
-    public ArticleDto getArticleDtoById(long articleId) {
-        Article article = articleRepository.findById(articleId).orElseThrow(NoSuchElementException::new);
+    public ArticleResponse getArticleDtoById(long articleId) {
+        Article article = getArticleById(articleId);
         return ArticleAssembler.writeDto(article);
     }
 
-    public Article save(ArticleDto articleDto) {
-        Article newArticle = ArticleAssembler.writeArticle(articleDto);
-        return articleRepository.save(newArticle);
+    public Article getArticleById(long articleId) {
+        return articleRepository.findById(articleId)
+                .orElseThrow(NoSuchElementException::new);
+    }
+
+    public Article save(ArticleRequest articleDto, User user) {
+        Article article = ArticleAssembler.writeArticle(articleDto, user);
+        return articleRepository.save(article);
     }
 
     @Transactional
-    public Article update(ArticleDto articleDto) {
-        Article updatedArticle = new Article(articleDto.getArticleId(), articleDto.getTitle(),
-                articleDto.getCoverUrl(), articleDto.getContents());
-        Article article = articleRepository.findById(updatedArticle.getArticleId())
-                .orElseThrow(NoSuchElementException::new);
+    public Article update(long articleId, ArticleRequest articleDto, User user) {
+        Article article = getArticleById(articleId);
+        article.checkCorrespondingAuthor(user);
+        Article updatedArticle = new Article(articleDto.getTitle(), articleDto.getCoverUrl(), articleDto.getContents(), user);
         return article.update(updatedArticle);
     }
 
-    public void deleteById(long articleId) {
+    public void deleteById(long articleId, User user) {
+        Article article = getArticleById(articleId);
+        article.checkCorrespondingAuthor(user);
         articleRepository.deleteById(articleId);
+    }
+
+    public List<CommentResponse> getAllComments(long articleId) {
+        Article article = getArticleById(articleId);
+        return CommentAssembler.writeDtos(article.getComments());
+    }
+
+    public Comment getCommentById(long commentId) {
+        return commentRepository.findById(commentId)
+                .orElseThrow(NoSuchElementException::new);
+    }
+
+    @Transactional
+    public Comment saveComment(long articleId, CommentRequest commentDto, User user) {
+        Article article = getArticleById(articleId);
+        Comment comment = CommentAssembler.writeComment(commentDto, user);
+        article.saveComment(comment);
+        return commentRepository.save(comment);
+    }
+
+    @Transactional
+    public Comment updateComment(long commentId, CommentRequest commentDto, User user) {
+        Comment comment = getCommentById(commentId);
+        comment.checkCorrespondingAuthor(user);
+        Comment updatedComment = CommentAssembler.writeComment(commentDto, user);
+        return comment.update(updatedComment);
+    }
+
+    public void deleteComment(long commentId, User user) {
+        Comment comment = getCommentById(commentId);
+        comment.checkCorrespondingAuthor(user);
+        commentRepository.deleteById(commentId);
     }
 }
